Codecov: disable Numba; ignore tests, experimental, and gpu_rtx (#673)
* analytics: test no name

* codecov: disable numba

* convolve 2d: dtype of array at initialization

* flake8

* run codecov once when push to PR

* ignore test polygonize when measuring codecov with numba disabled

* ignore experimental from codecov report

* exclude tests from codecov report

* retry

* codecov retry

+++ xrspatial/tests/test_zonal.py
@@ -250,7 +250,7 @@ def test_apply():
     def func(x):
         return 0
 
-    zones_val = np.zeros((3, 3), dtype=int)
+    zones_val = np.zeros((3, 3), dtype=np.int32)
     # define some zones
     zones_val[1] = 1
     zones_val[2] = 2
